Improve Kubernetes exception reporting
Move the Kubernetes object from a Slack field to a block, since
it's too long for a field. Remove the brackets around the kind,
since it seems reasonably clear without them. Add the kind to the
exception summary as well, and handle a few more formatting edge
cases.

KubernetesError was used in a couple of places that weren't
Kubernetes API errors. Switch them to either MissingObjectError or
a new DuplicateObjectError, and also improve the Slack formatting
of those exceptions.

Reorder parameters to follow a more natural naming order of kind,
namespace, name (from least specific to most specific), except for
the methods that take a name and a namespace, where we keep the
annoying Kubernetes API order of name, namespace to avoid confusion.