Handle groups without GIDs
Gafaelfawr allows the user to be a member of a group with no GID,
but this should be ignored for Nublado controller purposes since
we can't do anything useful with such groups. Filter them out when
constructing the lab and when constructing the user metadata portion
of the lab status.

Take advantage of the opportunity to refactor the method that
generates the NSS ConfigMap.