Add a new generic Kubernetes object watcher
Add a generic Kubernetes watcher class with proper typing for the
objects that it returns, built on the assumption that the Kubernetes
library will properly decode Kubernetes objects in the watch events.
For now, use it only to watch for pod status; the other uses of
watches in the Kubernetes storage layer will be replaced in upcoming
changes.

After some experimentation, this uses a different approach than the
callback approach currently used in the storage layer for watches.
Instead, the watch exposes an async interator with structured events
with appropriate typing, allowing any analysis to be done in the body
of the async for loop.