Add new generic Kubernetes storage layer
Add a couple of classes that provide generic Kubernetes storage
access, one to types where we only do create and read and another
for types where we also do list and delete. Support deleting the
old resource on conflicts for the latter type. Instantiate classes
for ConfigMap and Service management from those generic classes
and use them in the Kubernetes storage layer.