Add a new-style NamespaceStorage class
Replace the direct calls to the Kubernetes namespace API with a
new NamespaceStorage class in the new style. This class is entirely
separate from those based on KubernetesObjectDeleter even though
there is some duplication of code, since the namespace API is
fundamentally different (only a name, no namespace).

Preserve the more complex timeout handling for namespace deletion
since it is most likely to have timeout issues. This may be worth
implementing for all deleter objects, but that's a problem for a
future day.