Add new JobStorage and IngressStorage classes
Replace the Kubernetes job manipulation classes with a JobStorage
class that inherits from KubernetesObjectDeleter. Add
propagation_policy support to the standard delete method to support
its usage with jobs.

Replace the Kubernetes ingress calls with a new-style IngressStorage
class. This inherits the standard functionality from the generic
KubernetesObjectDeleter class, but adds a method to check whether an
ingress has an IP address assigned and a method to wait for an ingress
to appear and get an IP address.

This uncovered a bug that meant that the code waiting for the ingress
and job to be deleted was doing nothing due to a missing f in an
f-string. Fixing that in turn required some changes to the test suite,
since it was often not deleting the ingress until after the controller
was waiting for the ingress to go away.