Add Kubernetes conversion to LabResources
Tell the LabResources model how to generate a V1ResourceRequirements
object, and use that instead of repeating the same code in three
places. The fileserver code had a complex function to handle a
configuration that set partial resource requests or limits, but our
configuration rules that out because all limits and requests are
mandatory. For now, just drop all of that complexity; if we later
want to add support for partial resource configuration, we can re-add
that directly in the LabResources model.