Refactor prepuller storage layer
Model the new separation between builder objects, services, and
the Kubernetes storage layer with the prepuller, which is the
simplest. Add a new PrepullerBuilder that constructs the pod that
the prepuller uses, and have it use that and the PodStorage storage
layer directly, so it no longer needs K8sStorageClient.

Introduce a new MetadataStorage layer, which is resposible for
getting metadata about the running pod from the Kubernetes downward
API, and use it for namespace and object reference information for
the prepuller pods.

Stop using the namespace prefix for labs as the namespace for
prepuller pods and always use the namespace from the downward API.
Make the downward API information mandatory; there's really no
reason not to have it.

Stop adding Argo CD annotations to the prepuller pods. Since they
are children of the controller pod, they don't need any.