Move all lab object creation into a builder object
Move the builder object for the prepuller into a separate module
and move the lab builder into services.builder.lab. Rework how lab
creation is done to move all Kubernetes object construction into
the builder, which returns a dataclass containing all the objects.
Add a new lab-specific storage layer that's responsible for making
the calls to create those objects, plus some ancillary methods for
secret management.

Move the merging of secret data into the lab service, where it seems
to fit better, although there's no ideal location for it.

Make the Argo CD application to which lab objects are attached
configurable, rather than hard-coding it. Move a bunch of paths
inside the lab container into constants in preparation for making
them configurable.

Simplify the naming of volumes inside the pod specification. These
are all scoped to the pod, so there's no reason to add a username-nb
prefix to all of them. Also simplify the naming of files in the file
config map to use only the file name instead of the full path, since
the internal naming of the volumes already assumed that the file name
was unique. If we need to support files of the same name in multiple
locations in the future, we can cross that bridge then. Since they
already weren't working, take the readability win.

Ensure that the nublado.lsst.io/user annotation is set uniformly on
all lab objects.

Rename the helper functions for reading JSON test data to explicitly
read JSON data so that we can add one that reads plain files. Merge a
few home directory test cases and put them with the other lab tests.