Move fileserver object creation into a builder
Follow the new builder pattern for fileserver object creation and
move the Kubernetes object construction into FileserverBuilder.
Move construction of Kubernetes classes for volumes and volume mounts
into another small class that can be used by both the LabBuilder and
the FileserverBuilder. Move most of the storage layer for the
fileserver into a new storage class.

The storage object migration is not yet complete because of state
management and the calls to reconstruct state from Kubernetes objects.
That redesign will be handled in a subsequent redesign of the
separation between the manager and state classes to prepare for
Redis separation, after which the old Kubernetes storage layer can
be retired completely.

This surfaced a problem with the new watcher class. It was rounding
down timeouts instead of rounding them up, which caused functions to
be called with a zero timeout. Round up explicitly and also error out
if the timeout is set but is 0 or less.