Redo how disabled fileserver is handled
To make testing easier and allow enabling of the fileserver after
the FastAPI application has been started, unconditionally register
the fileserver routes but raise exceptions in the fileserver state
manager if it's not enabled. Change DisabledError to
NotConfiguredError to match the exception for the same purpose in
Gafaelfawr and configure it to translate into 404 errors for users.