Add support for homedir prefix and suffix
Allow the prefix for home directories (normally /home) and the
suffix (normally empty, but allowing the home directory for labs
to be a subdirectory of the user's normal home directory) to be
configured.

I picked this approach over adding template support because I don't
like having to document a template language in user documentation
and try to explain what variables are available and how to write a
template. In this case, I think the three options are reasonably
easy to understand and avoid the need for a full-blown templating
mini-language.