Install pytest-pretty and tweak test output
Use pytest-pretty to make the test output look a little prettier, and
update Python dependencies. Shorten the name of one test that may be
pushing the length of the verbose output. Disable verbose output by
default and instead set the CI environment variable, which forces rich
diffs of data structures.