Move lab state reconstruction into builder
Move the code to reconstruct lab state from objects on disk out of
the state manager and into the lab builder so that it lives side by
side with the code that creats those objects. In order to maintain
the separation between the builder and the storage layer, introduce
a couple of new dataclasses, one that holds all the names for user
objects and can be passed down to the storage layer, and one that
holds just the Kubernetes objects required for state reconstruction.

Tidy up the builder API by removing a few now-unnecessary methods
that were too low-level, and clean up more now-unnecessary methods
in the old Kubernetes storage layer.