Refactor lab state management
The production IDF cluster was seeing a steady stream of KeyError
exceptions trying to look up a user's lab in the _labs structure
in the LabStateManager. I believe this is due to two general class
of race conditions that might cause simultaneous modifications of
the state structure. This extensive refactoring attempts to address
each of them.

The first is two simultaneous operations for the same user. There
was some attempt to prevent this, but it had various race conditions
due to time-of-check/time-of-use bugs, made worse by the previous
behavior of deleting the user's lab state from the dictionary entirely
when the lab is deleted and thus creating a window for two lab
creations to race. The protection against simultaneous delete and
spawn operations was also lacking.

The second was a general class of races with background reconciliation.
The reconciliation thread attempted to exclude labs where an operation
was in progress, but there were time-of-check/time-of-use bugs with
that exclusion logic as well, also hampered by deleting the lab state
when the user's lab was deleted.

I attempted various solutions to this within the existing structure,
but the division of responsibilities between the LabManager and
LabStateManager classes added a great deal of interface complexity
and created code structure problems because of reconciliation living
in LabStateManager but lab spawning living in LabManager. I therefore
took the approach here of unifying LabManager and LabStateManager to
ensure one class was responsible for all overall state management.

Responsibility for ensuring that only one lab operation is in progress
at a time, and monitoring operations for exceptions or timeouts and
doing error reporting and cleanup, is now handled by a per-user
LabMonitor class that uses locking to ensure that a given user cannot
have multiple conflicting operations in progress. This allows supporting
multiple simultaneous deletes, all of which will block on the same
underlying delete operation. Spawn plus delete or multiple spawns will
now be rejected with 409 conflict responses.