Handle terminating and terminated status
Update the lab status enum to match the current Nublado controller,
including renaming starting to pending. Add support for terminated
status, which will be used for labs killed by the idle culler.
Update comments for the (soon to be implemented) behavior of the
Nublado controller.

Treat terminating status the same as terminated: tell the hub that
the lab has exited, even though it hasn't yet. We want to allow the
user to start spawning a new lab, and if they kick that off before
the lab deletion completes, we'll join the existing delete call now
that the controller supports multiple simultaneous deletes.