Allow delete to abort an in-progress spawn
Previously, we returned 409 if someone attempted to delete a spawn
that was still in progress. This interferes with cleanup that we
want to do in the REST Spawner if JupyterHub was restarted while
spawns are in progress. JupyterHub may lose the credentials for those
labs and need to stop them so that they can be restarted with
correct credentials, but the previous constraint meant that the stop
would fail until the spawn completed.

Use the new cancel method on the monitor to stop any spawn in progress
and then continue with a delete to clean up any remnant Kubernetes
resources. Send a failure event to the event stream of any in-progress
operation when the operation is cancelled.