Switch from pytest-pretty to pytest-sugar
pytest-pretty was a nice idea, but unfortunately due to the way that
it dumps local variables, it generates so much output on failure
tests that it was taking up to 30s for pytest to throw it away and
test failure reports were unusable. Switch to pytest-sugar instead
for some nice Unicode output but none of the other special features.