Delete not-running labs during reconciliation
When we encounter a lab in a terminated state during background
reconciliation (labs that were OOMKilled, idle-culled, etc.), delete
the lab. Do this carefully to avoid conflicts with a user operation
that might race with us, since it should get precedence. Similarly
clean up labs that are in a failed state.

Convert a pod status of Failed to a lab status of terminated, not
failed. Use failed to indicate missing Kubernetes resources or
similar errors, and terminated to mean that the pod has stopped for
whatever reason.