Convert to the SQR-075 monorepo layout
We want to combine the components of Nublado into a single repository
with one set of documentation. Start by converting the Nublado
controller to the new layout, renaming its Python module to controller
(since it won't be on PyPI, it doesn't really matter what we call it),
and using nox instead of tox to control the build since it works better
with the monorepo configuration.