Handle arguments to the test session
For now, only the controller has tests, so allow the path to
individual test files to be given with a controller prefix (so that
tab completion works), and remove that prefix when running pytest.
We will need to do something more complex in the future when more
than one component has its own tests.