Add Python 3.10 tests for Hub plugins
The JupyterHub image currently still uses Python 3.10, so we need
to test the authenticator and spawner modules with Python 3.10, but
the controller requires at least Python 3.11. Add a separate CI
run to test just the authenticator and spawner with a matrix of
Python versions that are supported by the Hub but not the controller.

Add a new nox environment that tests the authenticator and spawner
with the frozen dependencies for JupyterHub. Exclude the Ruff rule
to change timezone.utc to UTC for the spawner, since this was added
in Python 3.11, and change the code back to timezone.utc.