Remove pod state tracking
Previously, we attempted to track whether the pod for a user's lab
was present. After some practical experience, this seems pointless.
The important information is already captured in the lab state, and
if we really want to know if the pod exists (it's mostly not clear
that this matters), we can look directly in Kubernetes, since it would
only be for debugging purposes. Drop this tracking.