Convert fileserver creation to new template system
Model the template code after Gafaelfawr and create a FastAPI
template manager instead of directly creating a Jinja environment.

Rename the fileserver.timeout configuration option to idle_timeout
to more clearly distinguish it from creation_timeout, and make both
of those timeouts timedeltas instead of seconds. (Pydantic will
parse an integer as a number of seconds for timedelta fields.)

Move the formatter for the idle timeout in the template into a custom
formatter for Jinja rather than doing the formatting outside of Jinja.