Redo spawner form generation
Use the new templating system for generating the spawner form and
generate the form directly in the handler, since it is properly part
of the UI layer rather than the service layer.

Enhance the LabSizeDefinition config class and move some of the
formatting into the lab builder and the spawner HTML template, which
allows eliminating FormManager and the formdata method of SizeManager.