Partly overhaul the controller configuration
Review and flesh out documentation for all models in the controller
configuration. Sort the fields generally alphabetically.

Move the Safir settings into the main configuration, since they are
not specific to Safir and users have no reason to care about the
Safir abstraction level.

Refactor file server configuration to use a model hierarchy to
represent whether the file server is enabled, allowing proper full
validation of the file server configuration when it is enabled.
mypy does a great job of resolving the correct type after checking
the enabled field.

Change mode for volumes to readOnly, since mode was a two-item enum
that was essentially equivalent to a boolean.

Create an abstraction for Docker container specifications and use it
for both the file server and for init containers. This will allow
Mend Renovate to detect that this is a Docker container reference and
create PRs to update it.