Diagnose invalid lab size on creation
The model checking ensured that the requested lab size was a valid
member of the enum, but not that it was one of the configured
sizes. This could have resulted in a run-time KeyError if the user
sent an invalid size. Catch this error explicitly and return a 422
error instead.