Allow use of SI units for file server resources
The resource constraints on the file server reused the existing
ResourceQuantity model, which meant that memory limits and requests
had to be specified in bytes. Lift the bitmath utility function into
a new module for unit conversions and use it for both lab sizes and
for validation of ResourceQuantity, allowing use of SI and NIST
suffixes.

Switch to using bitmath's parse_string_unsafe, which handles the
lack of a B suffix, supports K instead of k, and otherwise more
closely matches the resource language used by Kubernetes.