Handle race condition reaping lab deletions
There was a race condition between the background task to reap
completed lab operations and lab deletions. Deletions are started
and then immediately waited for, which means they could be collected
by the handler between the point the background task sees they are
done and the point where it tries to reap them. It should be safe to
ignore this race in the background reaper task, since it's only point
is to reap tasks that no one else has waited for, so give that
exception a custom class and suppress it.