Add request timeouts to all Kubernetes calls
The default request timeout in kubernetes_asyncio if none is given
is five minutes, which is rather long. We were already setting
request timeouts for any watches. Set a shorter request timeout of
one minute for other operations so that we don't wait too long if
the Kubernetes control plane is not responding.

This timeout is currently hard-coded, but this change makes it
possible to make it configurable in the future.