Separate volumes and volume mounts
Previously, we were hoping to simplify the Kubernetes data model
by combining volume specifications and their mount points into one
data structure, under the assumption that each volume would only
be mounted in one place.

Unfortunately, this broke for several reasons: we do have use cases
where we want to mount the same volume on multiple paths, init
containers must share volume definitions with the main container but
do not share mounts, and file servers may eventually need the same
volumes but different mounts. Separate the list of volumes and the
list of mounts and tie them together with volume names, which are also
used in constructing the Kubernetes objects.

This enables support for PVCs in init containers nearly for free; it
just requires a bit of code in constructing the init container, which
is added here. While doing so, also add support for PVCs for file
servers, which is more complex since they have to be created and
deleted and do not have fixed names.

Name the PVCs after the volume name instead of assigning them numerical
names now that we have a volume name available.