Wait for file server pod creation properly
The Nublado controller was creating a Job and then immediately asking
Kubernetes for the child Pod, and then failing when it hadn't been
created yet even though it would show up in a moment. Wait for the
Ingress IP address assignment first, since it handles a missing
Ingress, and wait for the Pod second. Poll repeatedly for the Pod
creation until the timeout is exhausted. (This should use a watch, but
that's a bit more complex; that will come in a subsequent change.)