Overhaul Kubernetes timeout handling (again)
Add a new Timeout class that starts a timer when created and has
methods to get the elapsed time, error message, and number of seconds
left. Require a Timeout parameter to every Kubernetes storage layer
function, and plumb that machinery all the way up to the service layer,
using these Timeout objects for cumulative timeouts shared by a batch
of operations. Operations that aren't part of a batch with their own
configured timeout get a default timeout.

Add new configuration options for timeouts on deleting labs and file
servers.