Add separate exception for timeouts
Add a ControllerTimeoutError exception that is similar to TimeoutError
but with a better message and with support for posting to Slack. Have
the Timeout class take over calling asyncio.timeout, and add support to
it for translating the exception to ControllerTimeoutError and adding
the additional exception metadata.

This allows removal of the error method on timeouts, since they now
handle their own exception creation.