Remove clickable link to file server
Remove the clickable link to the user's file server at the top of
the instructions, since the file server only supports WebDAV and
following that link in a browser will not do anything useful.

Leave the link farther down after the instructions about how to
access it in case people want to use copy link location.