Add Python API documentation for plugins
Add Python API documentation in the dev section for the JupyterHub
plugins. Do a proofreading pass on that rendered documentation and
clean up some cross-references.

Unfortunately, there doesn't seem to be a way to reference one
package installed by Nublado from another package installed by
Nublado in a way that Sphinx understands well enough to generate
cross-references, so the mention of GafaelfawrAuthenticator in
the NubladoSpawner documentation is not a link.