Add internal API documentation for the controller
Generate Python API documentation for most of the controller,
following our normal practice of excluding the handler routes
since they don't have useful docstrings. This caught a bunch of
missing __all__ settings.

Enable autodoc_pydantic for Pydantic models and add templates
for those and for exceptions to try to improve the autogenerated
documentation.

Generating the documentation for RSPImage required not deferring
type annotations for RSPImageTag, since otherwise the semver
VersionInfo class was not resolved.

Sphinx generated weird errors for context_dependency and
config_dependency, the global variables holding the singleton
classes. Documenting those is not important, so they were dropped
from the list of objects to generate documentation for to resolve
those errors.