Add nodeSelector support to the prepuller
If nodeSelector is set for Nublado labs, apply the same node
selector when choosing nodes that are eligible for prepulling.
For node selector, just filter out the nodes from the prepuller
status entirely rather than listing them but showing why they
aren't eligible for prepulling. We'll use the latter strategy for
tolerations, since tolerations are more temporary.

Use the new function for formatting Kubernetes objects for
comparison purposes for the prepuller test objects.