Add tolerations support to the prepuller
If config.lab.tolerations are set, use them to select which nodes
to which to prepull images. Report the nodes that are excluded by
tolerations (but not the ones excluded by nodeSelector) in the
prepull status as not eligible.

This required some restructuring of the data stored by the image
service to retain more information about Kubernetes nodes.