Fix race conditions for labs and file servers
Both labs and file servers had race conditions with background
reconcile where reconcile might try to retrieve data while a spawn
was in progress, see an incomplete configuration, and decide to
delete the objects. The previous care with which the reoncile
method was written was not sufficient to avoid this.

To try to address this problem, add last modified timestamps to
the internal state that are updated whenever the lab or file server
is changed (created or deleted), and check those timestamps as well
as whether an operation is in progress when doing reconciliation.
If the state has changed since the collection of data for
reconciliation started, leave it alone and wait for the next pass.