Adjust how frozen dependencies are handled
Add mypy and, where needed, pydantic to the dev.in dependency
configuration rather than installing it in noxfile so that it's
properly version-pinned, ensuring that one doesn't start seeing
sudden mypy errors in the middle of development.

Add pytest-sugar to the test dependencies for inithome for
consistency.

Update Python dependencies.