Update to Python 3.12
Make the controller and inithome require Python 3.12, following our
normal policy of only supporting the latest version for code where
the release artifact is a Docker image. Add Python 3.12 to the support
and test matrix for the JupyterHub plugins. Suppress new warnings
from third-party packages with Python 3.12.

Set the development status of the components of Nublado to
production/stable, which is not strictly true yet but will be soon.