Set privileged for privileged init containers
Privileged init containers were being run as root but weren't being
marked as privileged. This may be required to do some types of setup
and was documented to be the case, and was being set with Nublado v3.
Add that flag.

Add a test of a non-privileged init container and providing multiple
init containers.