Show output from Sphinx linkcheck
Sphinx linkcheck doesn't present a summary of errors, which means
that the failures can be lost in the very long output, particularly
in GitHub actions where the failures aren't marked in color. Use a
tactic from other packages to check for the output file and
display it if the linkcheck command fails.