Rework the lab status models
The lab status routes were returning the full lab environment,
which isn't desirable since the environment may contain secrets.
They were also returning the complicated parameters we accept as
input to specify the image, rather than clearly returning the
image as chosen.

Fix this by splitting some of the models between input and output
models and renaming them to avoid the word User for anything other
than information about the actual user. The status now omits the
environment and returns the full reference for the running image,
rather than the complicated parameters passed in by the user.

This does mean we no longer have a good way of capturing the menu
options that people are picking, but the right way to do that,
should we ever care, is to add more logging.

Doing this removes the need to reconstruct the environment when
reconciling images, so all of that code can be deleted.