Separate prepuller API models from configuration
Rather than reusing the same models for the prepuller API responses
and the prepuller configuration, have the latter inherit from the
former and add the camel-case aliasing support. This ensures that
the API models do not define aliases, and thus fixes the OpenAPI
documentation and API responses to use snake-case, which we prefer
for all JSON APIs.