Always retry Kubernetes watches on 410 errors
Rather than special-casing the case where a resource version is
specified, always retry Kubernetes watches on 410 errors, not just
those with timeouts. Remove the delay before retrying the watch.

We were seeing this error every five minutes, admittedly in
situations where the namespace was potentially being deleted, but it
appears to be safe to immediately retry 410 errors.