Add support for Python 3.13
Switch the controller and inithome containers to Python 3.13. Add
Python 3.13 to the test matrix for the JupyterHub plugins. Break
client testing into a separate GitHub Action job and also test it
against a matrix of Python 3.12 and 3.13. Update the periodic CI
check for various things that were added to the main CI but not
updated there.

Update Python dependencies.