Convert all model fields to Annotated
Use the `Annotated` approach for all model fields, and convert some
validators to functional validators where it makes sense to do so.
Move model configuration to the start of the model and simplify a
few fields. Take advantage of the fact that dictionary keys can also
have `Annotated` types to put validators closer to the type.