Fix link checking issues
Sphinx appears to turn URLs in `Annotated` arguments into actual
links, which is new and annoying behavior. Add an example to the
exclusion list. Remove a now-unneeded paragraph that was generating
redirects, and change the setuptools-scm link. Only run a linkcheck
if there were changes to the documentation; running it for every
change of any type feels like a bit much.