Add support for blocking spawn with a quota flag
Gafaelfawr now supports a flag in the user's notebook quota saying
whether that user should be allowed to spawn a notebook. Implement
support for that flag in the Nublado controller. If it is set to false,
return an error message instead of a spawn form, and reject the actual
spawn with a 403 error.