Fix handling of timeouts in Kubernetes operations
When load testing Nublado, the controller raised uncaught exceptions
and returned HTTP 500 errors to JupyterHub due to Kubernetes operation
timeouts. This happened because the timeout parameter on the Kubernetes
API call caused `TimeoutError` to be raised, but the Kubernetes storage
layer was not translating that into the corresponding internal
`ControllerTimeoutError` exception that would be caught and handled.

Wrap all Kubernetes API operations in `Timeout.enforce`, which will
translate the `TimeoutError` exceptions properly so that they can be
handled.

Catch `ControllerTimeoutError` exceptionsn when getting the lab status
and optimistically return the current in-memory state, matching the
behavior when we receive Kubernetes API errors when checking the lab
state. JupyterHub makes this call constantly and we don't want to fail
the lab due to some transient timeout that will later recover.`