Do not query Kubernetes for pod status
Do not query Kubernetes for pod status when responding to status
requests, and instead assume the internal state (which is reconciled
periodically) is correct. The constant Kubernetes API requests for
`Pod` status seemed to be overwhelming the control plane when there
were a lot of pods running. Continue to ask Kubernetes directly
immediately before spawn.