Add tolerations to prepuller pods
When Nublado was configured to add tolerations to lab pods, those
tolerations were not also applied to prepuller pods. This meant that
prepuller pods were never run if the node pool targeted by Nublado
was tainted, such as when using a node pool restricted to only
Nublado labs. Add the tolerations to prepuller pods as well.