Fix state tracking when deleting completed labs
The Nublado controller background reconciliation thread looks for
labs that are no longer running and cleans up their resources. This
loop was calling an async function in the middle of the loop over
all labs, which meant that the dictionary over which it was looping
could be changed by a different task in the middle of the loop.
Instead, capture the current state first and loop over a snapshot of
the lab to lab state mapping.