Support per-user subdomains in the Nublado client
In the Nublado client, check the first redirect when accessing
JupyterHub or JupyterLab and see if it is a redirect to a different
host. If so, update the internal base URL of JupyterHub or JupyterLab
to the new base URL. This ensures that subsequent calls that cannot
be redirected (`POST` or `DELETE`, for instance) will use the correct
base URL and not fail due to a redirect or missing cookies.

Reorganize the Nublado client code to group together internal methods.
Reduce the debug logging for searching for XSRF cookies to only note
when a new XSRF cookie was found, instead of logging every step of
the process.