Simplify how client mocks are registered
Rework the code for registering mock routes with respx in the Nublado
client mock to make the code a bit easier to follow.

The `base_url` parameter to `mock_jupyter` is now used only for the
hostname and no longer supports arbitrary prefixes before the `/nb`
portion of the route, a feature that we were not using. This argument
will be changed in the future when we have a release of the client
that breaks backwards compatibility.