Fix NetworkPolicy for user labs
The `NetworkPolicy` created for labs essentially did nothing: it
tried to match JupyterHub or its proxy, which don't exist in the user
namespace, and then created an ingress rule with no restrictions.

Fix the `NetworkPolicy` to behave as intended: Apply rules to the user
lab and restrict access to only applications that are part of
JupyterHub.